Remove some unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325249 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 2f8da6f7e18c7c19436eff75997a83c4d8689495

+++ src/components/org/apache/jmeter/timers/ConstantTimer.java
@@ -82,7 +82,7 @@ public class ConstantTimer
      */
     public double getRange()
     {
-        return (double) 0;
+        return 0;
     }
 
     /**
